[Refactor] Move StarRocksSinkManagerV2 to stream-load-sdk so that other connectors can use it
